refactor: remove dead browser OAuth flow
- Remove --browser flag (server-side relay page was never built)
- Delete oauth.go and oauth_test.go
- Move ReadTokenFromStdin to device.go
- Device flow is the only interactive auth method
