+++ cmd/auth/login.go
@@ -12,12 +12,6 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// BrowserLoginFunc is the function used to perform browser-based login.
-// Tests may override this to avoid opening a real browser.
-var BrowserLoginFunc = func(host string, ios *iostreams.IOStreams) (string, error) {
-	return internalauth.BrowserLogin(host, ios)
-}
-
 // DeviceLoginFunc is the function used to perform device-code login.
 // Tests may override this to avoid real network calls.
 var DeviceLoginFunc = func(host string, ios *iostreams.IOStreams) (string, error) {
@@ -43,15 +37,14 @@ func NewLoginCmd(f *cmdutil.Factory) *cobra.Command {
 		Args:  cobra.NoArgs,
 		Long: `Authenticate with KeeperHub using the device code flow.
 Opens a browser to confirm a one-time code.
-Use --browser for direct browser-based OAuth (useful if device flow is unavailable).
 Use --with-token to read an API key from stdin for non-interactive automation.
 
 See also: kh auth status, kh auth logout`,
 		Example: `  # Log in (device code flow)
   kh auth login
 
-  # Log in via direct browser OAuth
-  kh auth login --browser`,
+  # Log in with an API key (non-interactive)
+  echo "kh_xxx" | kh auth login --with-token`,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			hosts, err := config.ReadHosts()
 			if err != nil {
@@ -67,13 +60,11 @@ See also: kh auth status, kh auth logout`,
 			envHost := os.Getenv("KH_HOST")
 			host := hosts.ActiveHost(flagHost, envHost)
 
-			useBrowser, _ := cmd.Flags().GetBool("browser")
 			withToken, _ := cmd.Flags().GetBool("with-token")
 
 			var token string
 
-			switch {
-			case withToken:
+			if withToken {
 				t, readErr := internalauth.ReadTokenFromStdin(f.IOStreams)
 				if readErr != nil {
 					return readErr
@@ -82,15 +73,7 @@ See also: kh auth status, kh auth logout`,
 					return fmt.Errorf("storing token: %w", err)
 				}
 				token = t
-
-			case useBrowser:
-				t, loginErr := BrowserLoginFunc(host, f.IOStreams)
-				if loginErr != nil {
-					return loginErr
-				}
-				token = t
-
-			default:
+			} else {
 				t, loginErr := DeviceLoginFunc(host, f.IOStreams)
 				if loginErr != nil {
 					return loginErr
@@ -104,7 +87,6 @@ See also: kh auth status, kh auth logout`,
 
 			info, err := FetchTokenInfoFunc(host, token)
 			if err != nil {
-				// Non-fatal: login succeeded but we can't fetch user details
 				fmt.Fprintf(f.IOStreams.Out, "Logged in to %s\n", host)
 				return nil
 			}
@@ -114,7 +96,6 @@ See also: kh auth status, kh auth logout`,
 		},
 	}
 
-	cmd.Flags().Bool("browser", false, "Use direct browser OAuth instead of device code flow")
 	cmd.Flags().Bool("with-token", false, "Read token from stdin")
 
 	return cmd
